+++ src/main/java/kitchenpos/application/TableServiceJpa.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.jpa;
+package kitchenpos.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
