+++ src/test/java/kitchenpos/application/OrderServiceJpaTest.java
@@ -1,8 +1,7 @@
-package kitchenpos.application.jpa;
+package kitchenpos.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 import kitchenpos.domain.entity.Menu;
